# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CamWatch — Multi-tenant IP camera monitoring API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List detection events
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: cameraId
          schema:
            type: integer
        - in: query
          name: type
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventList"
    post:
      operationId: createEvent
      tags: [events]
      summary: Ingest a new detection event (used by edge agents)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get event details
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete event
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Deleted

  # ─── ZONES ──────────────────────────────────────────────────────────────
  /zones:
    get:
      operationId: listZones
      tags: [zones]
      summary: List detection zones for the tenant
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: cameraId
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ZoneList"
    post:
      operationId: createZone
      tags: [zones]
      summary: Create a detection zone
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateZoneBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Zone"
  /zones/{id}:
    patch:
      operationId: updateZone
      tags: [zones]
      summary: Update detection zone
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateZoneBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Zone"
    delete:
      operationId: deleteZone
      tags: [zones]
      summary: Delete detection zone
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Deleted

  # ─── NOTIFICATIONS ──────────────────────────────────────────────────────
  /notifications:
    get:
      operationId: listNotificationChannels
      tags: [notifications]
      summary: List notification channels
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationChannelList"
    post:
      operationId: createNotificationChannel
      tags: [notifications]
      summary: Create notification channel
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNotificationChannelBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationChannel"
  /notifications/{id}:
    patch:
      operationId: updateNotificationChannel
      tags: [notifications]
      summary: Update notification channel
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNotificationChannelBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationChannel"
    delete:
      operationId: deleteNotificationChannel
      tags: [notifications]
      summary: Delete notification channel
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Deleted

  # ─── EDGE AGENTS ────────────────────────────────────────────────────────
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users in the tenant
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserList"
    post:
      operationId: inviteUser
      tags: [users]
      summary: Invite a new user to the tenant
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteUserBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user role
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: removeUser
      tags: [users]
      summary: Remove user from tenant
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Removed
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    Event:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        cameraId:
          type: integer
        cameraName:
          type: string
        type:
          type: string
          enum: [vehicle, car, motorcycle, truck, person, motion, line_cross, zone_enter, zone_exit, vehicle_stopped]
        confidence:
          type: number
        snapshotUrl:
          type: string
        metadata:
          type: object
          additionalProperties: true
        source:
          type: string
          enum: [cloud_events, edge_ai, onvif]
        detectedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, tenantId, cameraId, type, source, detectedAt, createdAt]
    EventList:
      type: array
      items:
        $ref: "#/components/schemas/Event"
    CreateEventBody:
      type: object
      properties:
        cameraId:
          type: integer
        type:
          type: string
          enum: [vehicle, car, motorcycle, truck, person, motion, line_cross, zone_enter, zone_exit, vehicle_stopped]
        confidence:
          type: number
        snapshotUrl:
          type: string
        metadata:
          type: object
          additionalProperties: true
        source:
          type: string
          enum: [cloud_events, edge_ai, onvif]
        detectedAt:
          type: string
          format: date-time
      required: [cameraId, type, source]

    # ─── ZONE ───────────────────────────────────────────────────────────
    Zone:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        cameraId:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [polygon, line, rectangle]
        points:
          type: array
          items:
            type: object
            properties:
              x:
                type: number
              y:
                type: number
            required: [x, y]
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, tenantId, cameraId, name, type, points, active, createdAt]
    ZoneList:
      type: array
      items:
        $ref: "#/components/schemas/Zone"
    CreateZoneBody:
      type: object
      properties:
        cameraId:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [polygon, line, rectangle]
        points:
          type: array
          items:
            type: object
            properties:
              x:
                type: number
              y:
                type: number
            required: [x, y]
        active:
          type: boolean
      required: [cameraId, name, type, points]
    UpdateZoneBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [polygon, line, rectangle]
        points:
          type: array
          items:
            type: object
            properties:
              x:
                type: number
              y:
                type: number
            required: [x, y]
        active:
          type: boolean

    # ─── NOTIFICATION CHANNEL ───────────────────────────────────────────
    NotificationChannel:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [telegram, whatsapp, email, webhook, push]
        config:
          type: object
          additionalProperties: true
        active:
          type: boolean
        triggerTypes:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required: [id, tenantId, name, type, config, active, triggerTypes, createdAt]
    NotificationChannelList:
      type: array
      items:
        $ref: "#/components/schemas/NotificationChannel"
    CreateNotificationChannelBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [telegram, whatsapp, email, webhook, push]
        config:
          type: object
          additionalProperties: true
        active:
          type: boolean
        triggerTypes:
          type: array
          items:
            type: string
      required: [name, type, config]
    UpdateNotificationChannelBody:
      type: object
      properties:
        name:
          type: string
        config:
          type: object
          additionalProperties: true
        active:
          type: boolean
        triggerTypes:
          type: array
          items:
            type: string

    # ─── EDGE AGENT ─────────────────────────────────────────────────────
...
    User:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [owner, admin, operator, viewer]
        createdAt:
          type: string
          format: date-time
      required: [id, tenantId, name, email, role, createdAt]
    UserList:
      type: array
      items:
        $ref: "#/components/schemas/User"
    InviteUserBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        role:
          type: string
          enum: [admin, operator, viewer]
      required: [name, email, password, role]
    UpdateUserBody:
      type: object
      properties:
        role:
          type: string
          enum: [admin, operator, viewer]
        name:
          type: string