# lib/api-zod/src/generated/types/createZoneBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CamWatch — Multi-tenant IP camera monitoring API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateZoneBodyPointsItem } from "./createZoneBodyPointsItem";
import type { CreateZoneBodyType } from "./createZoneBodyType";

export interface CreateZoneBody {
  cameraId: number;
  name: string;
  type: CreateZoneBodyType;
  points: CreateZoneBodyPointsItem[];
  active?: boolean;
}